requirements.txt
Django==2.2
djangorestframework==3.12.4
python-decouple==3.4
